requirements.txt
# Core framework
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1

# Database and env config
psycopg2-binary==2.9.11
python-dotenv==1.2.2

# Redis + Celery
redis==7.4.0
django-redis==6.0.0
celery==5.6.3

# Resume parsing
pypdf==6.9.2

# Salary model inference
joblib==1.5.3
pandas==3.0.0
scikit-learn==1.6.1
